Clojure solid icon and green face for repl
Use the solid clojure icon version for Clojure, matching other modes such as Emacs,
markdown, etc, rather than the clojure-line icon.

The clojure icon has greater visibility that the clojure-line icon, so it is
easier to see the major mode at a glance.

Change REPL face to green, indicating a running Clojure process and providing a
subtle distinction between source and repl buffers.  This matches the face
used to indicate a running process in modeline themes.