Prioritise regexps over file extension lookups
When looking for an icon for a given file name,
`all-the-icons-icon-for-file` uses two sources of information in the
following order:

1) `all-the-icons-extension-icon-alist`.
2) `all-the-icons-regexp-icon-alist`.

However, this might be incorrect in some cases when
`all-the-icons-regexp-icon-alist` is used to particularise cases. For
instance, with `bar_spec.rb` as input,
`all-the-icons-icon-for-file` produces:

```lisp
ELISP> (all-the-icons-icon-for-file "bar_spec.rb")
 #("" 0 1
   (face
    (:family "github-octicons" :height 1.2 :inherit all-the-icons-lred)
    font-lock-face
    (:family "github-octicons" :height 1.2 :inherit all-the-icons-lred)
    display
    (raise 0.0)
    rear-nonsticky t))
```

Which is a "generic" Ruby file icon, however the association list that
contains file name regexps has these elements:

```lisp
    ("_?test\\.rb$"        all-the-icons-fileicon "test-ruby" ...)
    ("_?test_helper\\.rb$" all-the-icons-fileicon "test-ruby" ...)
    ("_?spec\\.rb$"        all-the-icons-fileicon "test-ruby" ...)
    ("_?spec_helper\\.rb$" all-the-icons-fileicon "test-ruby" ...)
```

which help finding a better icon in this case. Unfortunately, 1) wins
and 2) is never explored.

This patch fixes this issue by giving the regex match a higher
priority than the file extension lookup. Also, to prevent the lookup
from stopping by falling into the "fallback" match in 2), I'm moving
the default to a separate variable.

With the patch applied, a better icon is returned in this case.

```lisp
ELISP> (all-the-icons-icon-for-file "bar_spec.rb")
 #("" 0 1
   (face
    (:family "file-icons" :height 1.2 :inherit all-the-icons-red)
    font-lock-face
    (:family "file-icons" :height 1.2 :inherit all-the-icons-red)
    display
    (raise 0.0)
    rear-nonsticky t))
```